finished breaking up token components into files
