+++ src/react-components/layout/CenteredModalWrapper.scss
@@ -1,0 +1,13 @@
+@use "../styles/theme.scss";
+
+:local(.fullscreen) {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0px;
+  left: 0px;
+}
+
+:local(.darken) {
+  background-color: rgba(0, 0, 0, 0.2);
+}
