+++ src/react-components/tokens/TokenList.js
@@ -12,6 +12,7 @@ import styles from "./Tokens.scss";
 export const TokenList = ({ tokens, onRevokeToken }) => {
   return (
     <div>
+      <TokenMenuHeader />
       <Row spaceBetween breakpointColumn="sm" topMargin="lg">
         <h2>
           <FormattedMessage id="empty-token.title2" defaultMessage="Token List" />
@@ -41,3 +42,18 @@ TokenList.propTypes = {
   tokens: PropTypes.array,
   onRevokeToken: PropTypes.func
 };
+
+// TODO move to TokenContainer when defining token state
+const TokenMenuHeader = () => (
+  <Column gap="xl">
+    <h1>
+      <FormattedMessage id="empty-token.title" defaultMessage="API Tokens" />
+    </h1>
+    <p>
+      <FormattedMessage id="tokens.create-tokens-description" defaultMessage="Create tokens to access the" />{" "}
+      <a href="#" rel="noreferrer noopener">
+        <FormattedMessage id="tokens.create-tokens-description-hubs-api" defaultMessage="Hubs API" />
+      </a>.
+    </p>
+  </Column>
+);
