+++ src/react-components/tokens/Tokens.scss
@@ -34,23 +34,6 @@ h2 {
   border: 2px solid theme.$red;
 }
 
-:local(.fullscreen) {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0px;
-  left: 0px;
-  // position: fixed;
-  // top: 50%;
-  // left: 50%;
-  // transform: translate(-50%, -50%);
-  // box-sizing: border-box;
-}
-
-:local(.darken) {
-  background-color: rgba(0, 0, 0, 0.2);
-}
-
 :local(.center-padding-sides) {
     width: 80%;
     max-width: 800px;
